Extract `withSlatePasting` into an extension
