+++ packages/slate-editor/src/extensions/paste-slate-content/PasteSlateContentExtension.ts
@@ -1,0 +1,12 @@
+import type { Extension } from '@prezly/slate-commons';
+
+import { withSlatePasting } from './lib';
+
+export const EXTENSION_ID = 'PasteSlateContentExtension';
+
+export function PasteSlateContentExtension(): Extension {
+    return {
+        id: EXTENSION_ID,
+        withOverrides: withSlatePasting,
+    };
+}
