+++ packages/slate-editor/src/extensions/paste-slate-content/index.ts
@@ -1,0 +1,1 @@
+export { PasteSlateContentExtension, EXTENSION_ID } from './PasteSlateContentExtension';
