+++ packages/slate-editor/src/extensions/paste-slate-content/lib/index.ts
@@ -1,0 +1,1 @@
+export { withSlatePasting } from './withSlatePasting';
