+++ packages/slate-editor/src/extensions/paste-slate-content/lib/withSlatePasting.ts
@@ -1,16 +1,9 @@
-/* eslint-disable no-param-reassign */
-
 import { decodeSlateFragment } from '@prezly/slate-commons';
 import { Editor } from 'slate';
 
-import { createDataTransfer } from '../lib';
-import { isFragment } from '../types';
+import { createDataTransfer } from '#lib';
 
-function withoutSlateFragment(data: DataTransfer): DataTransfer {
-    const types = data.types.filter((type) => type !== 'application/x-slate-fragment');
-    const dataMap = Object.fromEntries(types.map((type) => [type, data.getData(type)]));
-    return createDataTransfer(dataMap);
-}
+import { isFragment } from './isFragment';
 
 export function withSlatePasting<T extends Editor>(editor: T) {
     const { insertData } = editor;
@@ -25,7 +18,7 @@ export function withSlatePasting<T extends Editor>(editor: T) {
             if (isPrezlyFragment) {
                 Editor.insertFragment(editor, fragment);
             } else {
-                editor.insertData(withoutSlateFragment(data));
+                editor.insertData(withoutSlateFragmentAttribute(data));
             }
 
             return;
@@ -35,4 +28,10 @@ export function withSlatePasting<T extends Editor>(editor: T) {
     };
 
     return editor;
+}
+
+function withoutSlateFragmentAttribute(data: DataTransfer): DataTransfer {
+    const types = data.types.filter((type) => type !== 'application/x-slate-fragment');
+    const dataMap = Object.fromEntries(types.map((type) => [type, data.getData(type)]));
+    return createDataTransfer(dataMap);
 }
