+++ packages/slate-editor/src/lib/createDataTransfer.ts
