+++ packages/slate-editor/src/modules/editor/index.ts
@@ -1,7 +1,6 @@
 export { default as Editor } from './Editor';
 export { createEditor } from './createEditor';
 export {
-    createDataTransfer,
     createEmptyValue,
     createEmptyEditorValueAsString,
     createEditorValueWithCoverageAsString,
