+++ packages/slate-editor/src/modules/editor/plugins/index.ts
@@ -4,5 +4,4 @@ export { withNonEmptyValue } from './withNonEmptyValue';
 export type { RichBlocksAwareEditor } from './withRichBlocks';
 export { withRichBlocks } from './withRichBlocks';
 export { withRootElements } from './withRootElements';
-export { withSlatePasting } from './withSlatePasting';
 export { withVoids } from './withVoids';
