+++ packages/slate-editor/src/modules/editor/plugins/withDeserializeHtml/withDeserializeHtml.ts
@@ -5,9 +5,9 @@ import type { Extension } from '@prezly/slate-commons';
 import { EditorCommands } from '@prezly/slate-commons';
 import type { Editor } from 'slate';
 
-import { EventsEditor } from '#modules/events';
+import { createDataTransfer } from '#lib';
 
-import { createDataTransfer } from '../../lib';
+import { EventsEditor } from '#modules/events';
 
 import { deserializeHtml } from './deserializeHtml';
